refactor(lifecycle): remove evaluators and regeneration, add settle
