+++ src/lifecycle-constants.ts
@@ -2,11 +2,6 @@ export const TOTAL_MAX_STEPS = 200;
 export const INITIAL_MAX_STEPS = 80;
 export const STEP_TIMEOUT_MS = 120_000;
 export const MAX_UNKNOWN_ERRORS_PER_REQUEST = 2;
-export const MAX_REGENERATIONS_PER_REQUEST = 5;
-export const MAX_GUARD_RECOVERY_REGENERATIONS_PER_REQUEST = 2;
-export const MAX_LINT_REGENERATIONS_PER_REQUEST = 1;
-export const MAX_TOOL_RECOVERY_REGENERATIONS_PER_REQUEST = 2;
-export const MAX_REPEATED_FAILURE_REGENERATIONS_PER_REQUEST = 1;
 export const MAX_NUDGES_PER_GENERATION = 2;
 export const TOOL_TIMEOUT_MS = 10_000;
 export const CONSECUTIVE_GUARD_BLOCK_LIMIT = 5;
