+++ src/lifecycle-finalize.ts
@@ -57,8 +57,6 @@ export function phaseFinalize(ctx: RunContext): ChatResponse {
     write_calls: writeCalls,
     pre_write_discovery_calls: preWriteDiscoveryCalls,
     lifecycle_signal: ctx.result?.signal ?? null,
-    regeneration_count: ctx.regenerationCount,
-    regeneration_limit_hit: ctx.regenerationLimitHit,
     guard_blocked_count: guardStats.blocked,
     guard_flag_set_count: guardStats.flagSet,
     active_skill: ctx.promptUsage.activeSkillName ?? null,
