+++ src/lifecycle.test.ts
@@ -28,7 +28,7 @@ const phaseGenerate = mock(async (ctx: { result?: unknown }) => {
   ctx.result = { text: "Generated output", toolCalls: [], signal: "done" };
 });
 
-const phaseEvaluate = mock(
+const phaseSettle = mock(
   async (ctx: { session: { flags: { totalStepLimit?: number } }; result?: { text: string } }) => {
     expect(ctx.session.flags.totalStepLimit).toBe(12);
     expect(ctx.result?.text).toBe("Generated output");
@@ -70,7 +70,7 @@ describe("runLifecycle", () => {
       createRunAgent,
       phaseGenerate,
       shouldYieldNow: () => false,
-      phaseEvaluate,
+      phaseSettle,
       phaseFinalize,
     };
 
@@ -90,7 +90,7 @@ describe("runLifecycle", () => {
     expect(phasePrepare).toHaveBeenCalledTimes(1);
     expect(createRunAgent).toHaveBeenCalledTimes(1);
     expect(phaseGenerate).toHaveBeenCalledTimes(1);
-    expect(phaseEvaluate).toHaveBeenCalledTimes(1);
+    expect(phaseSettle).toHaveBeenCalledTimes(1);
     expect(phaseFinalize).toHaveBeenCalledTimes(1);
     expect(response).toEqual({ state: "done", model: "gpt-5-mini", output: "Generated output" });
     expect(debugEvents).toContain("lifecycle.workspace.sandbox");
