+++ src/test-utils.ts
@@ -468,7 +468,7 @@ export function createRunContext(overrides: Partial<RunContext> = {}): RunContex
       includedHistoryMessages: 0,
       totalHistoryMessages: 0,
     },
-    lifecycleState: { feedback: [] },
+    lifecycleState: {},
     observedTools: new Set(),
     modelCallCount: 1,
     inputTokensAccum: 0,
@@ -477,14 +477,6 @@ export function createRunContext(overrides: Partial<RunContext> = {}): RunContex
     streamingChars: 0,
     lastUsageEmitChars: 0,
     generationAttempt: 0,
-    regenerationCount: 0,
-    regenerationCounts: {
-      "guard-recovery": 0,
-      lint: 0,
-      "tool-recovery": 0,
-      "repeated-failure": 0,
-    },
-    regenerationLimitHit: false,
     errorStats: createErrorStats(),
     toolCallStartedAt: new Map(),
     toolOutputHandler: null,
